RESTORE UPPER-TO-DATE BUILD! YEY!! 💝🌱🥰
Should move from build 0202's source to build 10126's source! Once I thought that would be impossible, but this is a miracle happening right now!

This was developed after build 6883, a little before it turned into the build 12068. So, this is only a bit left to re-implement, which saved me lots of hours of work!

build 10126's spinoff CAPX is dated back in 'seg 08 jun 2015 21:09:50'